Do not catch callback exceptions
This is very problematic when trying to handle 'uncaughtException'
events (for example to log the error and then terminate the process).

Right now long-stack-traces will throw an empty string when catching a
callback error. This makes it impossible to get a reference to the
error in 'uncaughtException'.

This patch removes the catch statement, as this seems to be the sanest
way for long-stack-traces to react here.